renamed plugins dir to citest
